refactor(Ladder): move create line to ladder
